+++ src/OlaCounter.js
@@ -1,3 +1,4 @@
+
 const { registerDependencies } = require('mjml-validator');
 const { BodyComponent } = require('mjml-core');
 const tokens = require('./tokens');