+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - Unreleased
+### Added
+- New component `ola-item-list` [#8]
+- New component `ola-icon` [#8]
+- New component `ola-circle` [#8]
+- New component `ola-tag` [#8]
+- New component `ola-score` [#8]
+- New component `ola-progress-bar` [#8]
+- New components `ola-chart-column` and `ola-chart-column-value` [#8]
+
 ## [0.2.1] - 2020-09-17
 ### Fixed
 - CSS conflicts with `ola-button` component [#7]
@@ -46,7 +56,9 @@ First version
 
 [#5]: https://github.com/marketgoo/Ola-Emails/issues/5
 [#7]: https://github.com/marketgoo/Ola-Emails/issues/7
+[#8]: https://github.com/marketgoo/Ola-Emails/issues/8
 
+[0.3.0]: https://github.com/marketgoo/Ola-Emails/compare/v0.2.1...HEAD
 [0.2.1]: https://github.com/marketgoo/Ola-Emails/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.2...v0.1.3